+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * #1173 Add new click via `ClickOptions`  --  thanks to Dmytro Stekanov for PR #1226
 * #1259 Add new download via `DownloadOptions`  --  see PR #1260
 * #1272 Add check "href": `$("a").shouldHave(href("..."))`  --  see PR #1273
+* Add chrome option "--no-sandbox" (I hope it should make Chrome more stable)
 
 ## 5.14.2 (released 22.08.2020)
 * upgraded to WebDriverManager 4.2.0